+++ packages/cli/src/zosfiles/copy/ds/Ds.handler.ts
@@ -24,13 +24,15 @@ export default class DsHandler extends ZosFilesBaseHandler {
             "from-dataset": fromDataSet,
             enq: commandParameters.arguments.enq,
             replace: commandParameters.arguments.replace,
+            overwrite: commandParameters.arguments.overwrite,
             responseTimeout: commandParameters.arguments.responseTimeout,
             safeReplace: commandParameters.arguments.safeReplace,
             promptFn: this.promptForSafeReplace(commandParameters.response.console),
-            promptForIdenticalNamedMembers: this.promptForIdenticalNamedMembers(commandParameters.response.console)
+            promptForIdenticalNamedMembers: this.promptForIdenticalNamedMembers(commandParameters.response.console),
+            progress: commandParameters.response.progress,
         };
-
-        return Copy.dataSet(session, toDataSet, options);
+        const response = await Copy.dataSet(session, toDataSet, options);
+        return response;
     }
 
     private promptForSafeReplace(console: IHandlerResponseConsoleApi) {