+++ packages/zosfiles/src/methods/copy/doc/ICopyDatasetOptions.ts
@@ -1,3 +1,4 @@
+import { IHandlerProgressApi, ITaskWithStatus } from "@zowe/imperative";
 import { IDataSet } from "../../../doc/IDataSet";
 import { IZosFilesOptions } from "../../../doc/IZosFilesOptions";
 
@@ -36,6 +37,12 @@ export interface ICopyDatasetOptions extends IZosFilesOptions {
     replace?: boolean;
 
     /**
+     * Overwrite option
+     * @type {boolean}
+     */
+    overwrite?: boolean;
+
+    /**
      * Safe replace option
      * @type {boolean};
      */
@@ -50,7 +57,21 @@ export interface ICopyDatasetOptions extends IZosFilesOptions {
 
     /**
      * Prompt for duplicates
-     * @returns True if target data set members
+     * @returns `True` if the copy operation will overwrite an existing member; `false` otherwise
      */
     promptForIdenticalNamedMembers?: () => Promise<boolean>;
+
+    /**
+     * Task status object used by CLI handlers to create progress bars
+     * Optional
+     * @type {ITaskWithStatus}
+     */
+    task?: ITaskWithStatus;
+
+    /**
+     * Apis to create and destroy progress bars during the command.
+     * @type {IHandlerProgressApi}
+     * @memberof IHandlerResponseApi
+     */
+    progress?: IHandlerProgressApi;
 }